# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a video record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── PIPELINE ──
...
  /youtube/analytics:
    get:
      operationId: getYoutubeAnalytics
      tags: [youtube]
      summary: Get channel analytics overview
      parameters:
        - name: period
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /youtube/competitors:
    get:
      operationId: listCompetitors
      tags: [youtube]
      summary: List competitor channels
      responses:
        "200":
          description: Competitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"
    post:
      operationId: addCompetitor
      tags: [youtube]
      summary: Add a competitor channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorInput"
      responses:
        "201":
          description: Competitor added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
  /youtube/competitors/{id}:
    delete:
      operationId: removeCompetitor
      tags: [youtube]
      summary: Remove a competitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /youtube/keywords:
    get:
      operationId: listKeywords
      tags: [youtube]
      summary: List tracked keywords
      responses:
        "200":
          description: Keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: trackKeyword
      tags: [youtube]
      summary: Track a new keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "201":
          description: Keyword tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"

  # ── DRAFTS ──
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD SCHEMAS ──
...
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        youtubeId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        views:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        duration:
          type: ["integer", "null"]
        publishedAt:
          type: ["string", "null"]
        contentId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, title, type, status, views, likes, comments, createdAt]
    VideoInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        type:
          type: string
        contentId:
          type: ["integer", "null"]
      required: [title, type]
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        youtubeId:
          type: string
        thumbnailUrl:
          type: string
        views:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        duration:
          type: integer
        publishedAt:
          type: string

    # ── PIPELINE SCHEMAS ──
...
    AnalyticsOverview:
      type: object
      properties:
        period:
          type: string
        views:
          type: integer
        watchTime:
          type: number
        subscribers:
          type: integer
        revenue:
          type: number
        impressions:
          type: integer
        ctr:
          type: number
        avgViewDuration:
          type: number
        topVideos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        dailyStats:
          type: array
          items:
            $ref: "#/components/schemas/DailyStat"
      required: [period, views, watchTime, subscribers, revenue, impressions, ctr, avgViewDuration, topVideos, dailyStats]
    DailyStat:
      type: object
      properties:
        date:
          type: string
        views:
          type: integer
        watchTime:
          type: number
        revenue:
          type: number
        subscribers:
          type: integer
      required: [date, views, watchTime, revenue, subscribers]
    Competitor:
      type: object
      properties:
        id:
          type: integer
        channelName:
          type: string
        channelUrl:
          type: ["string", "null"]
        subscribers:
          type: integer
        totalVideos:
          type: integer
        avgViews:
          type: integer
        uploadFrequency:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, channelName, subscribers, totalVideos, avgViews, createdAt]
    CompetitorInput:
      type: object
      properties:
        channelName:
          type: string
          minLength: 2
        channelUrl:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
      required: [channelName]
    Keyword:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string
        volume:
          type: integer
        competition:
          type: string
        cpc:
          type: number
        trend:
          type: number
        tracked:
          type: boolean
        createdAt:
          type: string
      required: [id, keyword, volume, competition, cpc, trend, tracked, createdAt]
    KeywordInput:
      type: object
      properties:
        keyword:
          type: string
          minLength: 1
      required: [keyword]

    # ── DRAFT SCHEMAS ──
...