# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * YouTube AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  title: string;
  type: string;
  status: string;
  /** @nullable */
  youtubeId?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  views: number;
  likes: number;
  comments: number;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  contentId?: number | null;
  createdAt: string;
}